% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\docType{data}
\name{ann}
\alias{ann}
\title{Cell type annotations of `yan` datasets by Yan et al.}
\format{
An object of class \code{data.frame} with 90 rows and 1 columns.
}
\source{
\url{http://dx.doi.org/10.1038/nsmb.2660}

Each row corresponds to one cell of `yan` dataset
}
\usage{
ann
}
\description{
Cell type annotations of `yan` datasets by Yan et al.
}
\keyword{datasets}
