% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{BDSM}
\alias{BDSM}
\title{Calculate the bolock-diagnal matrix B
min_{B>=0, B*1=1, F'*F=I}  ||B - A||_1 + r*||B||^2 + 2*lambda*trace(F'*L*F)}
\usage{
BDSM(S, C)
}
\arguments{
\item{S}{the calculated initial similarity matrix S}

\item{C}{the estimated number of clusters C}
}
\value{
B block-diagonal matrix

y clustering results
}
\description{
Calculate the bolock-diagnal matrix B
min_{B>=0, B*1=1, F'*F=I}  ||B - A||_1 + r*||B||^2 + 2*lambda*trace(F'*L*F)
}
\examples{
gfData <- GenesFilter(yan)
res_SimS <- SimS(gfData)
C <- EstClusters(res_SimS$drData,res_SimS$S)
BDSM(res_SimS$S,C)

}
