% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFunctions.R
\name{EucDist}
\alias{EucDist}
\title{Solve the problem: ||A-B||^2 = ||A||^2 + ||B||^2 - 2*A'*B}
\usage{
EucDist(A, B)
}
\arguments{
\item{A}{matrix or vector}

\item{B}{matrix or vector}
}
\value{
d matrix or vector
}
\description{
Solve the problem: ||A-B||^2 = ||A||^2 + ||B||^2 - 2*A'*B
}
